chore: move qrCode job creation into JobController
