+++ src/server/modules/job/index.ts
@@ -1,0 +1,1 @@
+export { JobController, JobController as default } from './JobController'
